запустилось, отладка

+++ src/main/java/task_manager/testing/service/TaskAdminService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface TaskAdminService{
 
-    TaskDto createTaskForUser (TaskDto taskDto, Long UserId);
-    String deleteUserTask(TaskDto taskDto, Long UserId);
-    TaskDto updateUserTask(TaskDto taskDto, Long UserId);
+    TaskDto createTaskForUser (TaskDto taskDto, Integer UserId);
+    String deleteUserTask(TaskDto taskDto, Integer UserId);
+    TaskDto updateUserTask(TaskDto taskDto, Integer UserId);
     List<TaskDto> showAllTask();
 
 }
